==> 02 Building Energy Benchmarking/notebooks/0. Property Scrape.ipynb <==
# Property Scrape

import pandas as pd
import numpy as np

import requests
from bs4 import BeautifulSoup

from selenium import webdriver
from selenium.webdriver.common.keys import Keys
import time

import os
chromedriver = "./chromedriver"
os.environ["webdriver.chrome.driver"] = chromedriver

driver = webdriver.Chrome(chromedriver)

public_df = pd.read_csv('../data/raw/wcmsp-185933.csv')
public_df.info()

private_df = pd.read_csv('../data/raw/wcmsp-185935.csv')

private_df.head()

# Clean the address strings to the required format for Hennepin County Property Info search

def clean_addresses(df):
    df['search_address'].replace('Avenue', 'ave', inplace=True, regex=True)
    df['search_address'].replace('Ave\.', 'ave', inplace=True, regex=True)
    df['search_address'].replace('Ave', 'ave', inplace=True, regex=True)
    df['search_address'].replace('NE', 'n e', inplace=True, regex=True)
    df['search_address'].replace('SE', 's e', inplace=True, regex=True)
    df['search_address'].replace('Southeast', 's e', inplace=True, regex=True)
    df['search_address'].replace('Northeast', 'n e', inplace=True, regex=True)
    df['search_address'].replace('Southwest', 's w', inplace=True, regex=True)
    df['search_address'].replace('Curve', 'cur', inplace=True, regex=True)
    df['search_address'].replace('Parkway', 'pkwy', inplace=True, regex=True)
    df['search_address'].replace('Terrace', 'ter', inplace=True, regex=True)
    df['search_address'].replace('East', 'e', inplace=True, regex=True)
    df['search_address'].replace('S\.', 's', inplace=True, regex=True)
    df['search_address'].replace('South', 's', inplace=True, regex=True)
    df['search_address'].replace('Boulevard', 'blvd', inplace=True, regex=True)
    df['search_address'].replace('Curve', 'cur', inplace=True, regex=True)
    df['search_address'].replace('Drive', 'dr', inplace=True, regex=True)
    df['search_address'].replace('Trail', 'tr', inplace=True, regex=True)
    df['search_address'].replace('North', 'n', inplace=True, regex=True)
    df['search_address'].replace('Circle', 'cir', inplace=True, regex=True)
    df['search_address'].replace('Highway', 'hwy', inplace=True, regex=True)
    df['search_address'].replace('Road', 'rd', inplace=True, regex=True)
    df['search_address'].replace('Court', 'ct', inplace=True, regex=True)
    df['search_address'].replace('Lane', 'ln', inplace=True, regex=True)
    df['search_address'].replace('St\.', 'st', inplace=True, regex=True)
    df['search_address'].replace('Street', 'st', inplace=True, regex=True)
    df['search_address'].replace('street', 'st', inplace=True, regex=True)
    df['search_address'].replace('steet', 'st', inplace=True, regex=True)
    df['search_address'].replace('Northwest', 'n w', inplace=True, regex=True)
    df['search_address'].replace('West', 'w', inplace=True, regex=True)
    df['search_address'].replace('County', 'co', inplace=True, regex=True)
    df['search_address'].replace('Number', 'no', inplace=True, regex=True)
    df['search_address'].replace('Sixth', '6th', inplace=True, regex=True)

private_df['search_address'] = private_df['Address']
clean_addresses(private_df)
private_df.search_address

public_df['search_address'] = public_df['Address']
clean_addresses(public_df)
public_df.search_address

import re
import string

def selenium_search_url(driver, search_url, address):
    driver.get(search_url)

    address_num = address.split(" ")[0]
    address_street = address.split(" ")[1:]
    address_box = driver.find_element_by_id("INPUT1")
    address_box.send_keys(address_num)
    street_box = driver.find_element_by_id("INPUT3")
    street_box.send_keys(address_street)

    submit_button = driver.find_element_by_id("INPUT4")
    submit_button.click()
    
    property_url = driver.current_url
    print(property_url)

    content = (driver.page_source).encode('ascii', 'ignore')
    return content

def create_soup(content):
    soup = BeautifulSoup(content,"lxml")
    return soup

def search_multiple_properties(address, trs, data):
    for tr in trs:
        match = re.search(r'(pid\=)(\d*)(\"\>)', str(tr))
        if match:
            print(match.group(2))
            url = "http://www16.co.hennepin.mn.us/pins/pidresult.jsp?pid=" + match.group(2)
            driver.get(url)
            content = driver.page_source
            soup = create_soup(content)
            parse_property_page(address, soup, data) 
    return data

def parse_property_page(address, soup, data):
    new_record = {}
    
    if "Direct Debit program" in soup.find(id='TABLE_DETAILS').find_all('tr')[6].find_all('td')[0].text.strip():
        property_id = int(soup.find(id='TABLE_DETAILS').find_all('tr')[7].find_all('td')[1].text.strip().replace('-',''))
        new_record['address'] = address 
        new_record['property_id'] = property_id 
        new_record['construction_year'] = soup.find(id='TABLE_DETAILS').find_all('tr')[10].find_all('td')[4].text.strip() 
        new_record['owner_name'] = soup.find(id='TABLE_DETAILS').find_all('tr')[15].find_all('td')[1].text.strip()

        if "NO SALE INFORMATION ON FILE" in soup.find(id='TABLE_DETAILS').find_all('tr')[24].find_all('td')[0].text.strip():
            new_record['est_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[39].find_all('td')[1].text.strip()
            new_record['tax_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[40].find_all('td')[1].text.strip() 
            new_record['total_net_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[42].find_all('td')[1].text.strip() 
            new_record['total_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[45].find_all('td')[1].text.strip()
        else:
            new_record['sale_date'] = soup.find(id='TABLE_DETAILS').find_all('tr')[24].find_all('td')[1].text.strip()
            new_record['sale_price'] = soup.find(id='TABLE_DETAILS').find_all('tr')[25].find_all('td')[1].text.strip()
            new_record['transaction_type'] = soup.find(id='TABLE_DETAILS').find_all('tr')[26].find_all('td')[1].text.strip()
            #positions with NO sales section
            new_record['est_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[41].find_all('td')[1].text.strip()
            new_record['tax_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[42].find_all('td')[1].text.strip() 
            new_record['total_net_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[44].find_all('td')[1].text.strip() 
            new_record['total_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[47].find_all('td')[1].text.strip()
    
    elif "State Copy to be used when filing for the" in soup.find(id='TABLE_DETAILS').find_all('tr')[4].find_all('td')[0].text.strip():
        property_id = int(soup.find(id='TABLE_DETAILS').find_all('tr')[8].find_all('td')[1].text.strip().replace('-',''))
        new_record['address'] = address 
        new_record['property_id'] = property_id 
        new_record['construction_year'] = soup.find(id='TABLE_DETAILS').find_all('tr')[11].find_all('td')[4].text.strip() 
        new_record['owner_name'] = soup.find(id='TABLE_DETAILS').find_all('tr')[16].find_all('td')[1].text.strip()
        
        if "NO SALE INFORMATION ON FILE" in soup.find(id='TABLE_DETAILS').find_all('tr')[25].find_all('td')[0].text.strip():
            new_record['est_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[40].find_all('td')[1].text.strip()
            new_record['tax_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[41].find_all('td')[1].text.strip() 
            new_record['total_net_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[43].find_all('td')[1].text.strip() 
            new_record['total_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[46].find_all('td')[1].text.strip()
        else:
            new_record['sale_date'] = soup.find(id='TABLE_DETAILS').find_all('tr')[25].find_all('td')[1].text.strip()
            new_record['sale_price'] = soup.find(id='TABLE_DETAILS').find_all('tr')[26].find_all('td')[1].text.strip()
            new_record['transaction_type'] = soup.find(id='TABLE_DETAILS').find_all('tr')[27].find_all('td')[1].text.strip()
            #positions with sales section
            new_record['est_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[42].find_all('td')[1].text.strip()
            new_record['tax_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[43].find_all('td')[1].text.strip() 
            new_record['total_net_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[45].find_all('td')[1].text.strip() 
            new_record['total_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[48].find_all('td')[1].text.strip()
    else:  
        property_id = int(soup.find(id='TABLE_DETAILS').find_all('tr')[6].find_all('td')[1].text.strip().replace('-',''))
        new_record['address'] = address 
        new_record['property_id'] = property_id 
        new_record['construction_year'] = soup.find(id='TABLE_DETAILS').find_all('tr')[9].find_all('td')[4].text.strip() 
        new_record['owner_name'] = soup.find(id='TABLE_DETAILS').find_all('tr')[14].find_all('td')[1].text.strip()
        
        if "NO SALE INFORMATION ON FILE" in soup.find(id='TABLE_DETAILS').find_all('tr')[23].find_all('td')[0].text.strip():
            new_record['est_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[38].find_all('td')[1].text.strip()
            new_record['tax_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[39].find_all('td')[1].text.strip() 
            new_record['total_net_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[41].find_all('td')[1].text.strip() 
            new_record['total_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[44].find_all('td')[1].text.strip()
        else:
            new_record['sale_date'] = soup.find(id='TABLE_DETAILS').find_all('tr')[23].find_all('td')[1].text.strip()
            new_record['sale_price'] = soup.find(id='TABLE_DETAILS').find_all('tr')[24].find_all('td')[1].text.strip()
            new_record['transaction_type'] = soup.find(id='TABLE_DETAILS').find_all('tr')[25].find_all('td')[1].text.strip()
            #positions with NO sales section
            new_record['est_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[40].find_all('td')[1].text.strip()
            new_record['tax_market_value'] = soup.find(id='TABLE_DETAILS').find_all('tr')[41].find_all('td')[1].text.strip() 
            new_record['total_net_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[43].find_all('td')[1].text.strip() 
            new_record['total_tax'] = soup.find(id='TABLE_DETAILS').find_all('tr')[46].find_all('td')[1].text.strip()          
    
    new_record['land_market_value'] = soup.find(id='Table_Detail').find_all('tr')[3].find_all('td')[1].text.strip() 
    new_record['building_market_value'] = soup.find(id='Table_Detail').find_all('tr')[4].find_all('td')[1].text.strip() 
    new_record['total_market_value'] = soup.find(id='Table_Detail').find_all('tr')[6].find_all('td')[1].text.strip()
    new_record['homestead_mrkt_val_excl'] = soup.find(id='Table_Detail').find_all('tr')[9].find_all('td')[1].text.strip()
    new_record['property_type'] = soup.find(id='Table_Detail').find_all('tr')[11].find_all('td')[1].text.strip()
    new_record['homestead_status'] = soup.find(id='Table_Detail').find_all('tr')[12].find_all('td')[1].text.strip()

    data[property_id] = new_record
    return data

def run_scraper(df):
    chromedriver = "./chromedriver"
    driver = webdriver.Chrome(chromedriver)
    search_url="http://www16.co.hennepin.mn.us/pins/addrsrch.jsp"

    data = {}
    for index, row in df.iterrows():
        try:
            content = selenium_search_url(driver, search_url, row['search_address'])
            soup = create_soup(content)

            if soup.find(id='TABLE_DETAILS') and soup.find(id='TABLE_DETAILS').find_all('tr')[6].find_all('td')[1].text:
                print("Found " + row['Address'] + ": " + soup.find(id='TABLE_DETAILS').find_all('tr')[6].find_all('td')[1].text)
                parse_property_page(row['Address'], soup, data)
            elif soup.find(class_="notfound", string="No records found"):
                print("No records found for " +  row['Address'])
            elif soup.find("span", string="Click on property ID for details"):
                print(row['Address'] + " has multiple properties")
                select = driver.find_element_by_name("ps")
                select.send_keys('100')
                submit_button = driver.find_element_by_id("INPUT4")
                submit_button.click()

                content = (driver.page_source).encode('ascii', 'ignore')
                soup = create_soup(content)
                trs = soup.find(id='TABLE1').find_all('tr')
                search_multiple_properties(row['Address'], trs, data)
            else:
                print("Something went wrong " + row['Address'])
        except Exception:
            print("Exception on " +  row['Address'])
            pass
    return data
    

public_property_info = run_scraper(public_df)

import pickle

with open('0201_public_prop_dict.pkl', 'wb') as picklefile:
    pickle.dump(public_property_info, picklefile)

private_property_info = run_scraper(private_df)


#41 of 258 private properties were not found
import pickle

with open('0201_private_prop_dict.pkl', 'wb') as picklefile:
    pickle.dump(private_property_info, picklefile)

clean_addresses(public_df)
public_df.sample(5)

public_property_info = run_scraper(public_df)

private_df.head()
