==> 02 Building Energy Benchmarking/notebooks/1. Prepare Parcel Data.ipynb <==
# Python 2 & 3 Compatibility
from __future__ import print_function, division

# Necessary imports
import pandas as pd
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import patsy
import seaborn as sns
from seaborn import plt
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import RidgeCV

%matplotlib inline

from dbfread import DBF

# This file is too large to track in this Github repo

table = DBF('file.dbf', load=True, raw=True)

cols = table.field_names
cols

# create a dataframe

df = pd.DataFrame(table.records)

df.sample(5)

df.shape

# Turn the table from binary to a utf-8 encoded string

def decode_binary(df, cols):
    decode = lambda x : x.decode('utf-8')
    for col in cols:
        df[col] = df[col].apply(decode)
        

decode_binary(df, cols)
df.head()

# Remove white space from around the entries

def strip_entries(df, cols):
    strip = lambda x : x.strip()
    for col in cols:
        df[col] = df[col].apply(strip)

# Remove the county code and dash from the property IDs

def remove_dash(df, cols):
    x_dash = lambda x : x.split('-')[1] if '-' in x else x
    for col in cols:
        df[col] = df[col].apply(x_dash)

strip_entries(df, cols)

remove_dash(df, cols)

df.head()

# Remove empty strings

def replace_empty_values(df):
    df_cl = df.replace(r'\s+', np.nan, regex=True)
    df_cl = df.replace('', np.nan)
    return df_cl

df = replace_empty_values(df)

df.head()

# convert sale date to date object
import time

def convert_date(df):
    convert = lambda x : time.strptime(x, "%Y%m%d") if type(x) == str else np.nan
    df['sale_date_object'].apply(convert)
    
df['sale_date_object'] = df['SALE_DATE']

convert_date(df)

df['SALE_VALUE'].replace('[\$,)]','', inplace=True, regex=True)
df['SALE_VALUE'] = df['SALE_VALUE'].astype(float, copy=False)

df['YEAR_BUILT'] = df['YEAR_BUILT'].astype(int, copy=False)

df['FIN_SQ_FT'] = df['FIN_SQ_FT'].astype(int, copy=False)
df['EMV_LAND'] = df['EMV_LAND'].astype(float, copy=False)
df['EMV_BLDG'] = df['EMV_BLDG'].astype(float, copy=False)
df['EMV_TOTAL'] = df['EMV_TOTAL'].astype(float, copy=False)
df['TOTAL_TAX'] = df['TOTAL_TAX'].astype(float, copy=False)
df['TAX_CAPAC'] = df['TAX_CAPAC'].astype(float, copy=False)
df['SCHOOL_DST'] = df['SCHOOL_DST'].astype(float, copy=False)

df['cost_per_sqft'] = df['SALE_VALUE'] / df['FIN_SQ_FT']

df.head()

df_mpls = df[df['CITY'] == "MINNEAPOLIS"]

len(df_mpls)

# Extracted only the columns I want to use for this analysis

sub_cols = ['PIN', 'USE1_DESC','EMV_TOTAL', 'TAX_CAPAC', 'TOTAL_TAX', 'FIN_SQ_FT', 'YEAR_BUILT']
df_subset = df_mpls[sub_cols]

df_subset.info()

df_subset.head()

with open('../data/processed/0202_mpls_parcel_data_abr.pkl', 'wb') as picklefile:
    pickle.dump(df_subset, picklefile)
