==> 02 Building Energy Benchmarking/notebooks/2. Join Data.ipynb <==
# Python 2 & 3 Compatibility
from __future__ import print_function, division

# Necessary imports
import pandas as pd
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import patsy
import seaborn as sns
from seaborn import plt
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import RidgeCV

%matplotlib inline


import pickle

# 1. Data Import

# Open the prepped parcel dataframe

with open("../data/processed/0202_mpls_parcel_data_abr.pkl", 'rb') as picklefile: 
    parcel_df = pickle.load(picklefile)
parcel_df.head()

parcel_df.info()

# Open the private ordinance csv

private_df = pd.read_csv('../data/raw/wcmsp-185935.csv')
cols = ['address', 'postal_code', 'energy_star_score', 'primary_property_type',
       'building_floor_area', 'parking_floor_area',
       'year_built', 'total_ghg_emissions',
       'site_eui', 'weather_normalized_site_eui',
       'source_eui', 'weather_normalized_source_eui',
       'water_use', 'unnamed']

# Rename the columns
private_df.columns = cols
private_df.head()

priv_df = private_df[['address','energy_star_score',
       'building_floor_area', 'parking_floor_area','total_ghg_emissions',
       'site_eui', 'weather_normalized_site_eui',
       'source_eui', 'weather_normalized_source_eui',
       'water_use']]

priv_df

# Open the prepped property info dataframe

with open("../data/processed/0201_prop_info_data_abr.pkl", 'rb') as picklefile: 
    prop_info_df = pickle.load(picklefile)
prop_info_df.head()

# Convert the property_id from and into to a string for joining

prop_info_df['property_id_str'] = prop_info_df['property_id'].astype(str)


prop_info_df.info()

# Join the private ordinance data and the property info data

private_prop_info_df = pd.merge(priv_df, prop_info_df, how='inner', left_on='address', right_on='address')

private_prop_info_df.info()

# Join the private and property info data with the parcel data

join_df = pd.merge(private_prop_info_df, parcel_df, how='inner', left_on='property_id_str', right_on='PIN')

join_df.info()

# Drop remaining object fields

join_df.drop(['property_id_str','address', 'PIN'], axis=1, inplace=True)

join_df.info()

# Add calculated field

join_df['est_val_per_sqft'] = join_df['EMV_TOTAL'] / join_df['building_floor_area']

# Remove all records where the energy store score is 0

join_df2 = join_df[join_df['energy_star_score'] != 0]

join_df2.info()

join_df2.USE1_DESC.value_counts()

# Strip white space from the USE1_DESC entries

def strip_entries(df):
    strip = lambda x : x.strip()
    df['USE1_DESC'] = df['USE1_DESC'].apply(strip)

strip_entries(join_df2)
join_df2['USE1_DESC'].value_counts()

# Remove all residential entries
# TODO: Identify why they ended up the dataset?

join_df3 = join_df2[(join_df2['USE1_DESC'] != "Condominium")]
join_df4 = join_df3[(join_df3['USE1_DESC'] != "Residential")]
join_df5 = join_df4[(join_df4['USE1_DESC'] != "Apartment")]
df_non_res = join_df5[(join_df5['USE1_DESC'] != "Townhouse")]

df_non_res.info()

# Set misc random USE1_DESC entries to 'Other'

def set_use_to_other(df):
    df['USE1_DESC'].replace('Common Area (No Value)','Other', inplace=True)
    df['USE1_DESC'].replace('Low Income > 3 Units','Other', inplace=True)
    df['USE1_DESC'].replace('Double Bungalow','Other', inplace=True)

set_use_to_other(df_non_res)

df_non_res.USE1_DESC.value_counts()

with open('../data/processed/0202_all_data.pkl', 'wb') as picklefile:
    pickle.dump(df_non_res, picklefile)
