==> 02 Building Energy Benchmarking/notebooks/3. Prepare Dummy Variables.ipynb <==
# Python 2 & 3 Compatibility
from __future__ import print_function, division

# Necessary imports
import pandas as pd
import numpy as np
import statsmodels.api as sm
import statsmodels.formula.api as smf
import patsy
import seaborn as sns
from seaborn import plt
from sklearn.linear_model import LinearRegression
from sklearn.linear_model import RidgeCV

%matplotlib inline

import pickle

# Open the joined data

with open("../data/processed/0202_all_data.pkl", 'rb') as picklefile: 
    df = pickle.load(picklefile)

df.head()

df.info()

df.USE1_DESC.value_counts()

# Create dummy variable from USE1_DESC

X=patsy.dmatrix('USE1_DESC',data=df,return_type='dataframe')
X.head()

# Append the X matrix to your df matrix

df2 = df.join(X)

df2.head()

df2.drop('USE1_DESC', axis=1, inplace=True)

df2.info()

with open('../data/processed/0202_all_data_wdummy.pkl', 'wb') as picklefile:
    pickle.dump(df2, picklefile)
